podcast-generator/requirements.txt
langchain==0.1.17 
openai==1.26.0
python-dotenv==1.0.1
streamlit==1.34.0
langchain-openai==0.1.7
faiss-cpu==1.8.0
PyPDF2==3.0.1
elevenlabs==1.2.2